src/service/pyproject.toml
[tool.poetry]
name = "service"
version = "0.1.0"
description = ""
authors = ["Zoltan Kohalmy <zoltankohalmy@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
kafka-python = "^2.0.1"
sqlalchemy = "^1.3.18"
psycopg2-binary = "^2.8.5"
python-dateutil = "^2.8.1"

[tool.poetry.dev-dependencies]
pytest = "^5.4.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
